Fix echo_orient alias reads in MilkDrop expressions
